Start isolating the yaml parsing and problem building from the solver
